cleanup and remove deprecated async mode
* replace get_task_method with process_message
* rename NSQReader -> Reader; move init stuff around
